+++ 2libft/ft_putendl_fd.c
@@ -6,7 +6,7 @@
 /*   By: ochristm <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/08 18:06:34 by ochristm          #+#    #+#             */
-/*   Updated: 2019/09/08 18:06:48 by ochristm         ###   ########.fr       */
+/*   Updated: 2019/09/21 15:28:39 by ochristm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	ft_putendl_fd(char const *s, int fd)
 {
-	int i;
+	size_t	i;
 
 	i = 0;
 	if (s)
